Move location to utils
