+++ Cacom/src/grammar.lalrpop
@@ -3,8 +3,10 @@
 // https://github.com/kondziu
 
 use crate::ast::{StmtType, Opcode, ExprType, Stmt, Expr, Location};
+use crate::utils::Location;
 use std::collections::VecDeque;
 use std::str::FromStr;
+
 
 grammar;
